… smart-maintenance-saas/docs/COMPREHENSIVE_SYSTEM_ANALYSIS_REPORT.md
… smart-maintenance-saas/docs/COVERAGE_IMPROVEMENT_PLAN.md